Refactor. Moved shader_program class into shader file. Moved grid_manager class into grid file. Moved some grid_info definition stuff into ui_state file so header files work again.
