+++ grid.hpp
@@ -2,11 +2,8 @@
 
 #include <Eigen/Core>
 #include "cube.hpp"
+#include "ui_state.hpp"
 
-struct Coordinate {
-	int i;
-	int j;
-};
 
 class Chunk {
 public:
@@ -80,18 +77,6 @@ public:
 };
 
 //--------------------------------------------------------------------------------
-struct Grid_Info {
-	int rows;
-	int columns;
-
-	int origin_row; 
-	int origin_column;
-	int number_of_alive_cells;
-
-	int iteration;
-};
-
-//--------------------------------------------------------------------------------
 class Grid_Render_Data {
 public:
 	Grid_Render_Data();
@@ -130,3 +115,25 @@ public:
 	std::vector<Chunk> chunks;
 };
 
+//--------------------------------------------------------------------------------
+struct Grid_Execution_State {
+	bool is_running = false;
+	bool run_manual_next_iteration = false;
+	float time_since_last_iteration = 0.0f;
+	float grid_speed = 1.0f;
+};
+
+//--------------------------------------------------------------------------------
+class Grid_Manager {
+public:
+	Grid_Manager();
+
+	void initialise();
+
+	void update(double dt, Grid_UI_Controls_Info grid_ui_controls_info);
+
+	void create_new_grid();
+
+	Grid* grid;
+	Grid_Execution_State grid_execution_state;
+};
