+++ imgui.ini
@@ -14,7 +14,7 @@ Size=347,143
 Collapsed=0
 
 [Window][Dear ImGui Demo]
-Pos=1025,108
+Pos=123,115
 Size=550,746
 Collapsed=0
