+++ renderer.hpp
@@ -12,8 +12,7 @@
 #include <glm/gtc/matrix_transform.hpp>
 #include <glm/gtc/type_ptr.hpp>
 
-#include "shader_program.hpp"
-#include "texture.hpp"
+#include "shader.hpp"
 
 #include "state_render_data.hpp"
 #include "ui_state.hpp"
