+++ sancious_shaders.vcxproj
@@ -33,7 +33,6 @@
     <ClCompile Include="main.cpp" />
     <ClCompile Include="read.cpp" />
     <ClCompile Include="shader.cpp" />
-    <ClCompile Include="shader_program.cpp" />
     <ClCompile Include="state.cpp" />
     <ClCompile Include="texture.cpp" />
     <ClCompile Include="cube.cpp" />
@@ -55,7 +54,6 @@
 	<ClInclude Include="world.hpp" />
     <ClInclude Include="read.hpp" />
     <ClInclude Include="shader.hpp" />
-    <ClInclude Include="shader_program.hpp" />
     <ClInclude Include="stb_image.h" />
     <ClInclude Include="state.hpp" />
     <ClInclude Include="texture.hpp" />
