+++ state.cpp
@@ -20,61 +20,6 @@ void Timer::update() {
 	m_last_frame_time = m_current_frame_time;
 }
 
-Grid_Manager::Grid_Manager()
-: grid_execution_state({})
-{
-	ZoneScoped;
-
-	create_new_grid();
-};
-
-//--------------------------------------------------------------------------------
-void Grid_Manager::create_new_grid() {
-	ZoneScoped;
-	
-	grid_execution_state = {};
-	grid = new Grid();
-}
-
-void Grid_Manager::update(double dt, Grid_UI_Controls_Info ui_info) {
-	switch (ui_info.button_type) {
-		case GRID_NO_BUTTON_PRESSED:
-			break;
-		case GRID_RESET_BUTTON_PRESSED:
-			create_new_grid();
-			return;
-			// if you remove the return somehow later, dont forget a break statement here :)
-			// break;
-		case GRID_NEXT_ITERATION_BUTTON_PRESSED:
-			// only set run_manual_next_iteration if we are stopped
-			if (!grid_execution_state.is_running) {
-				grid_execution_state.run_manual_next_iteration = true;
-			}
-			break;
-		case GRID_START_STOP_BUTTON_PRESSED:
-			grid_execution_state.is_running = !grid_execution_state.is_running;
-			grid_execution_state.time_since_last_iteration = 0.0f;
-			break;
-		default:
-			break;
-	}
-	grid_execution_state.grid_speed = ui_info.grid_speed_slider_value;
-
-	if (grid_execution_state.is_running) {
-		//assert(grid_execution_state.grid_speed > 0.0f);
-		grid_execution_state.time_since_last_iteration += (float) dt;
-		float threshold = 1.0f / grid_execution_state.grid_speed;
-		if (grid_execution_state.time_since_last_iteration >= threshold) {
-			grid->next_iteration();
-			grid_execution_state.time_since_last_iteration = 0.0f;
-		}
-	} else {
-		if (grid_execution_state.run_manual_next_iteration) {
-			grid_execution_state.run_manual_next_iteration = false;
-			grid->next_iteration();
-		}
-	}
-}
 
 //--------------------------------------------------------------------------------
 State::State() : window(nullptr), timer(nullptr), ui_state(nullptr), renderer(nullptr),
@@ -108,8 +53,6 @@ void State::update() {
 	//World_Render_Data render_data = world->create_render_data();
 	// this sets up a new IMGUI-Frame! But we call the imgui render function only in our renderer! We always have to call ui_state->update() before the imgui render function,otherwise imgui didnt start a new frame!
 	ui_state->update(render_data->grid_render_data->grid_info);
-
-	
 }
 
 //--------------------------------------------------------------------------------
@@ -119,7 +62,6 @@ void State::initialise(GLFWwindow* w) {
 	renderer->initialise(window);
 	world->initialise(window);
 	ui_state->initialise(window);
-
 }
 
 //--------------------------------------------------------------------------------
