+++ state_render_data.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "world.hpp"
+
 //--------------------------------------------------------------------------------
 class World_Render_Data {
 public:
