+++ ui_state.hpp
@@ -5,8 +5,6 @@
 #include "imgui_impl_glfw.h"
 #include "imgui_impl_opengl3.h"
 
-#include "grid.hpp"
-
 enum Grid_UI_Control_Button_Events {
 	GRID_NO_BUTTON_PRESSED,
 	GRID_RESET_BUTTON_PRESSED,
@@ -18,6 +16,19 @@ struct Grid_UI_Controls_Info {
 	Grid_UI_Control_Button_Events button_type = GRID_NO_BUTTON_PRESSED;
 	float grid_speed_slider_value = 100.0f;
 };
+
+//--------------------------------------------------------------------------------
+struct Grid_Info {
+	int rows;
+	int columns;
+
+	int origin_row; 
+	int origin_column;
+	int number_of_alive_cells;
+
+	int iteration;
+};
+
 
 //--------------------------------------------------------------------------------
 class UI_State {
